Moved TokenRegex attribute to other folder
