+++ RobotPlusPlus.Compiler/Tokenizing/Token.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using RobotPlusPlus.Attributes;
 using RobotPlusPlus.Utility;
 
 namespace RobotPlusPlus.Tokenizing
